backend/pyproject.toml
[project]
name = "legalgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.11",
    "fastapi>=0.128.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]
